order-service/requirements.txt
asgiref==3.8.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.0.0
constantly==23.10.4
cryptography==42.0.8
daphne==4.0.0
Django==5.0.6
django-etc==1.4.0
django-filter==23.3
django-jalali==6.2.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==22.10.0
jalali_core==1.0.0
jdatetime==5.0.0
msgpack==1.0.8
pika==1.3.2
pilkit==3.0
Pillow==9.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
redis==5.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
zipp==3.19.2
zope.interface==6.4.post2
